+++ eFacturador/src/main/java/com/braintech/eFacturador/dto/inventario/InStockActualDTO.java
@@ -13,5 +13,5 @@ public class InStockActualDTO {
   private String productoNombre;
   private Integer almacenId;
   private String lote;
-  private Double cantidad;
+  private Integer cantidad;
 }
